requirements.txt
requests
beautifulsoup4
opencv-python
flask
ultralytics
python-telegram-bot
